Modified some url mappings

+++ parking-app/parking-app/src/main/java/com/be/parking_app/dto/PaymentMethodDTO.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 public class PaymentMethodDTO {
     private Integer paymentMethodId;
     private String cardType;
-    private Integer cardNumber;
+    private Long cardNumber;
     private Integer expiryMonth;
     private Integer expiryYear;
     private Integer securityCode;
